do a function check for error_get_last
git-svn-id: https://plugins.svn.wordpress.org/postman-smtp/trunk@1086229 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ postman.php
@@ -22,15 +22,18 @@ if (! isset ( $_SESSION )) {
 
 register_shutdown_function ( 'handleErrors' );
 function handleErrors() {
-	$last_error = error_get_last ();
-	$t = $last_error['type'];
-	$logger = new PostmanLogger ( 'postman.php' );
-	if (! is_null ( $last_error ) && ($t & (E_ERROR | E_PARSE)) && preg_match ( "/postman/i", $last_error ['file'] )) {
-		// if there has been a fatal error
-		$message = sprintf ( '%s in %s on line %d', $last_error ['message'], $last_error ['file'], $last_error ['line'] );
-		printf ( '<h2>Bad, Postman!</h2> <p><b><tt>X-(</b></tt></p> <p>Look at this mess:</p><code>%s</code>', $message );
-	} else {
-		$logger->debug ( 'Normal exit' );
+	// error_get_last is only in PHP 5.2 and newer
+	if (function_exists ( 'error_get_last' )) {
+		$last_error = error_get_last ();
+		$t = $last_error ['type'];
+		$logger = new PostmanLogger ( 'postman.php' );
+		if (! is_null ( $last_error ) && ($t & (E_ERROR | E_PARSE)) && preg_match ( "/postman/i", $last_error ['file'] )) {
+			// if there has been a fatal error
+			$message = sprintf ( '%s in %s on line %d', $last_error ['message'], $last_error ['file'], $last_error ['line'] );
+			printf ( '<h2>Bad, Postman!</h2> <p><b><tt>X-(</b></tt></p> <p>Look at this mess:</p><code>%s</code>', $message );
+		} else {
+			$logger->debug ( 'Normal exit' );
+		}
 	}
 }
